Added missing shortsword file.
The program loads all the weapon files into an array based on information in the master file. When rolling, we look at the weapon type selected by the user and find the index of that weapon in the weaponList array. That index is used to find the correct fumble max value and correct section of the weapon chart array.